+++ web/views/builder/country-flags.js
@@ -1,0 +1,69 @@
+// Country-flag mutual exclusion: picking "Global" clears individual countries
+// and picking any country clears Global. The four functions are referenced
+// from inline `onchange=` attributes generated by renderCountryFlagsHTML, so
+// builder.js exposes them on window during init().
+//
+// Edit-modal variants are pure DOM. Story-form variants additionally trigger
+// a debounced preview regenerate, which is the only injected dependency.
+
+const EDIT_FLAG_IDS = [
+    'editFlagCroatia', 'editFlagCzechia', 'editFlagGermany', 'editFlagHungary',
+    'editFlagIceland', 'editFlagItaly', 'editFlagPortugal', 'editFlagSlovakia',
+    'editFlagSlovenia', 'editFlagSpain', 'editFlagUK', 'editFlagFrance',
+];
+
+const STORY_FLAG_COUNTRIES = [
+    'croatia', 'czechia', 'germany', 'hungary', 'iceland', 'italy',
+    'portugal', 'slovakia', 'slovenia', 'spain', 'uk', 'france',
+];
+
+/**
+ * @param {{ onStoryChange: () => void }} deps
+ */
+export function createCountryFlagHandlers({ onStoryChange }) {
+    function clearEditGlobalIfCountrySelected() {
+        const anyCountrySelected = EDIT_FLAG_IDS.some((id) => document.getElementById(id)?.checked);
+        if (anyCountrySelected) {
+            const globalEl = document.getElementById('editFlagGlobal');
+            if (globalEl) globalEl.checked = false;
+        }
+    }
+
+    function clearEditCountriesIfGlobalSelected() {
+        const globalEl = document.getElementById('editFlagGlobal');
+        if (!globalEl?.checked) return;
+        EDIT_FLAG_IDS.forEach((id) => {
+            const el = document.getElementById(id);
+            if (el) el.checked = false;
+        });
+    }
+
+    function clearStoryGlobalIfCountrySelected(storyId) {
+        const anyCountrySelected = STORY_FLAG_COUNTRIES.some((country) =>
+            document.getElementById(`story-flag-${country}-${storyId}`)?.checked
+        );
+        if (anyCountrySelected) {
+            const globalEl = document.getElementById(`story-flag-global-${storyId}`);
+            if (globalEl) globalEl.checked = false;
+        }
+        onStoryChange();
+    }
+
+    function clearStoryCountriesIfGlobalSelected(storyId) {
+        const globalEl = document.getElementById(`story-flag-global-${storyId}`);
+        if (globalEl?.checked) {
+            STORY_FLAG_COUNTRIES.forEach((country) => {
+                const el = document.getElementById(`story-flag-${country}-${storyId}`);
+                if (el) el.checked = false;
+            });
+        }
+        onStoryChange();
+    }
+
+    return {
+        clearEditGlobalIfCountrySelected,
+        clearEditCountriesIfGlobalSelected,
+        clearStoryGlobalIfCountrySelected,
+        clearStoryCountriesIfGlobalSelected,
+    };
+}
